Format & lint src files. Rename main.js -> index.js
