+++ src/soundexRef/ref.1.js
@@ -1,40 +1,40 @@
 const ref = {
-        'ء': 0,
-        'ؤ': 0,
-        'ئ': 0,
-        'أ': 0,
-        'إ': 0,
-        'ا': 0,
-        'آ': 0,
-        'ه': 0,
-        'ع': 0,
-        'غ': 0,
-        'خ': 0,
-        'ح': 0,
-        'ق': 1,
-        'ك': 1,
-        'ج': 2,
-        'ش': 2,
-        'ي': 2,
-        'ى': 2,
-        'ض': 2,
-        'ص': 3,
-        'ز': 3,
-        'س': 3,
-        'ت': 4,
-        'ة': 4,
-        'ط': 4,
-        'د': 4,
-        'ث':5,
-        'ظ':5,
-        'ذ':5,
-        'ر': 6,
-        'ل': 6,
-        'ن': 6,
-        'ف': 7,
-        'و': 7,
-        'ب': 7,
-        'م': 7
-    };
+  ء: 0,
+  ؤ: 0,
+  ئ: 0,
+  أ: 0,
+  إ: 0,
+  ا: 0,
+  آ: 0,
+  ه: 0,
+  ع: 0,
+  غ: 0,
+  خ: 0,
+  ح: 0,
+  ق: 1,
+  ك: 1,
+  ج: 2,
+  ش: 2,
+  ي: 2,
+  ى: 2,
+  ض: 2,
+  ص: 3,
+  ز: 3,
+  س: 3,
+  ت: 4,
+  ة: 4,
+  ط: 4,
+  د: 4,
+  ث: 5,
+  ظ: 5,
+  ذ: 5,
+  ر: 6,
+  ل: 6,
+  ن: 6,
+  ف: 7,
+  و: 7,
+  ب: 7,
+  م: 7
+};
 
-    export default ref; 
+export default ref;
