+++ src/soundexRef/ref.2.js
@@ -1,36 +1,36 @@
 const ref = {
-    'ا': 0,
-    'أ': 0,
-    'إ': 0,
-    'آ': 0,
-    'ح': 0,
-    'خ': 0,
-    'ه': 0,
-    'ع': 0,
-    'غ': 0,
-    'ش': 0,
-    'و': 0,
-    'ي': 0,
-    'ف': 1,
-    'ب': 1,
-    'ج': 2,
-    'ز': 2,
-    'س': 2,
-    'ص': 2,
-    'ظ': 2,
-    'ق': 2,
-    'ك': 2,
-    'ت': 3,
-    'ث': 3,
-    'ة': 3,
-    'د': 3,
-    'ذ': 3,
-    'ض': 3,
-    'ط': 3,
-    'ل': 4,
-    'م': 5,
-    'ن': 5,
-    'ر': 6
+  ا: 0,
+  أ: 0,
+  إ: 0,
+  آ: 0,
+  ح: 0,
+  خ: 0,
+  ه: 0,
+  ع: 0,
+  غ: 0,
+  ش: 0,
+  و: 0,
+  ي: 0,
+  ف: 1,
+  ب: 1,
+  ج: 2,
+  ز: 2,
+  س: 2,
+  ص: 2,
+  ظ: 2,
+  ق: 2,
+  ك: 2,
+  ت: 3,
+  ث: 3,
+  ة: 3,
+  د: 3,
+  ذ: 3,
+  ض: 3,
+  ط: 3,
+  ل: 4,
+  م: 5,
+  ن: 5,
+  ر: 6
 };
 
 export default ref;
